Extract chunking into separate Chunker object
