+++ comm/src/main/scala/coop/rchain/comm/transport/TcpTransportLayer.scala
@@ -121,31 +121,6 @@ class TcpTransportLayer(port: Int, cert: String, key: String, maxMessageSize: In
   ): Task[CommErr[Option[Protocol]]] =
     withClient(peer, enforce)(f).attempt.map(processResponse(peer, _))
 
-  def chunkIt(blob: Blob): Task[Iterator[Chunk]] =
-    Task.delay {
-      val raw      = blob.packet.content.toByteArray
-      val kb500    = 1024 * 500
-      val compress = raw.length > kb500
-      val content  = if (compress) raw.compress else raw
-
-      def header: Chunk =
-        Chunk().withHeader(
-          ChunkHeader()
-            .withCompressed(compress)
-            .withContentLength(raw.length)
-            .withSender(ProtocolHelper.node(blob.sender))
-            .withTypeId(blob.packet.typeId)
-        )
-      val buffer    = 2 * 1024 // 2 kbytes for protobuf related stuff
-      val chunkSize = maxMessageSize - buffer
-      def data: Iterator[Chunk] =
-        content.sliding(chunkSize, chunkSize).map { data =>
-          Chunk().withData(ChunkData().withContentData(ProtocolHelper.toProtocolBytes(data)))
-        }
-
-      Iterator(header) ++ data
-    }
-
   def stream(peers: Seq[PeerNode], blob: Blob): Task[Unit] =
     streamObservable.stream(peers.toList, blob) *> log.info(s"stream to $peers blob")
 
@@ -232,7 +207,7 @@ class TcpTransportLayer(port: Int, cert: String, key: String, maxMessageSize: In
         case Right(packet) =>
           withClient(peer, enforce = false) { stub =>
             val blob = Blob(sender, packet)
-            stub.stream(Observable.fromIterator(chunkIt(blob)))
+            stub.stream(Observable.fromIterator(Chunker.chunkIt(blob, maxMessageSize)))
           }.attempt
             .flatMap {
               case Left(error) => log.error(s"Error while streaming packet, error: $error")
